+++ src/resources/test_data/common/common_data.json
@@ -14,6 +14,15 @@
   "advanced_filter_label": "Advanced filter",
   "result_count_heading": " results",
   "selected_checkboxes_hint_label": "selected",
+  "active_filters_label": "Active filters",
+  "no_results_heading": "There are no matching results",
+  "no_results_guidance_text": "Improve your search results by:",
+  "no_results_bullet_points": [
+    "removing filters",
+    "double-checking your spelling",
+    "using fewer keywords",
+    "searching for something less specific"
+  ],
   "No_Matching_Search_Results": {
     "no_matching_search_result_header_label": "There are no matching results",
     "no_matching_search_result_sub_header_label": "Improve your search results by:",
