+++ src/resources/test_data/iras/reviewResearch/receiveAmendments/search_modifications_page_data.json
@@ -30,11 +30,11 @@
     "modification_type_label": "Modification type",
     "modification_type_hint_text": "selected",
     "modification_type_checkbox": ["Minor modifications", "Modification of an important detail"],
+    "participating_nation_label": "Participating nation",
     "short_project_title_label": "Short project title",
     "short_project_title_hint_text": "Enter the name of the short project title",
     "sponsor_organisation_label": "Sponsor organisation",
     "sponsor_organisation_hint_text": "Type the name of the organisation or scroll through the list",
-    "active_filters_label": "Active filters",
     "Invalid_Date_Modification_Submitted_Error": {
       "date_modification_submitted_to_day_text": "The date should be a valid date and in future"
     },
